rag-contract-assistant-main/requirements.txt
streamlit
sentence-transformers
faiss-cpu
pdfplumber
python-docx
pypdf
huggingface_hub
requests
numpy
pandas
